ensure timestamp is utc
